fix(nms): Rename master portal to host portal
