+++ nms/packages/magmalte/app/components/__tests__/DashboardAlertTableTest.js
@@ -51,7 +51,7 @@ const mockAlertSt: Array<prom_firing_alert> = [
       alertname: 'TestAlert1',
       instance: '192.168.0.124:2112',
       job: 'myapp',
-      monitor: 'master',
+      monitor: 'host',
       networkID: 'test',
       severity: 'critical',
     },
@@ -73,7 +73,7 @@ const mockAlertSt: Array<prom_firing_alert> = [
       alertname: 'TestAlert2',
       instance: '192.168.0.124:2112',
       job: 'myapp',
-      monitor: 'master',
+      monitor: 'host',
       networkID: 'test',
       severity: 'major',
     },
@@ -95,7 +95,7 @@ const mockAlertSt: Array<prom_firing_alert> = [
       alertname: 'TestAlert3',
       instance: '192.168.0.124:2112',
       job: 'myapp',
-      monitor: 'master',
+      monitor: 'host',
       networkID: 'test',
       severity: 'minor',
     },
@@ -117,7 +117,7 @@ const mockAlertSt: Array<prom_firing_alert> = [
       alertname: 'TestAlert4',
       instance: '192.168.0.124:2112',
       job: 'myapp',
-      monitor: 'master',
+      monitor: 'host',
       networkID: 'test',
       severity: 'normal',
     },
