+++ nms/packages/magmalte/config/paths.js
@@ -25,7 +25,7 @@ const resolveApp = (relativePath: string) =>
 module.exports = {
   appIndexJs: resolveApp('app/main.js'),
   loginJs: resolveApp('app/login.js'),
-  masterJs: resolveApp('app/master.js'),
+  hostJs: resolveApp('app/host.js'),
   appSrc: resolveApp('app'),
   distPath: resolveApp('static/dist'),
   packagesDir: resolveApp('../../packages'),
