+++ nms/packages/magmalte/package.json
@@ -26,19 +26,21 @@
     "@babel/register": "^7.13.16",
     "@babel/runtime": "^7.14.0",
     "@date-io/moment": "1.3.13",
-    "@fbcnms/alarms": "^0.3.1",
-    "@fbcnms/auth": "^0.1.6",
+    "@fbcnms/alarms": "^1.0.5",
+    "@fbcnms/auth": "^0.2.0",
     "@fbcnms/babel-register": "^0.1.2",
-    "@fbcnms/express-middleware": "^0.1.5",
-    "@fbcnms/platform-server": "^0.3.3",
+    "@fbcnms/express-middleware": "^0.2.0",
+    "@fbcnms/platform-server": "^0.4.1",
     "@fbcnms/projects": "^0.1.0",
-    "@fbcnms/sequelize-models": "^0.1.13",
+    "@fbcnms/sequelize-models": "^0.2.0",
     "@fbcnms/strings": "^0.1.0",
     "@fbcnms/types": "^0.1.11",
-    "@fbcnms/ui": "^0.2.0",
+    "@fbcnms/ui": "^2.0.1",
     "@fbcnms/util": "^0.1.0",
+    "@material-table/core": "^4.3.36",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.0.0",
+    "@material-ui/lab": "^4.0.0-alpha.60",
     "@material-ui/pickers": "^3.2.10",
     "@material-ui/styles": "^4.0.0",
     "chart.js": "^2.7.3",
@@ -53,7 +55,9 @@
     "material-ui-dropzone": "^3.5.0",
     "pg": "^8.5.1",
     "pug": "^3.0.1",
+    "react": "^16.8.6",
     "react-chartjs-2": "^2.7.4",
+    "react-dom": "^16.8.6",
     "react-json-tree": "^0.11.2",
     "react-json-view": "^1.19.1",
     "react-router-dom": "5.1.2",
