+++ nms/packages/magmalte/scripts/dev_setup.sh
@@ -16,7 +16,7 @@ set -e
 # Add the test admin user
 docker-compose exec magmalte yarn migrate
 docker-compose exec magmalte yarn setAdminPassword magma-test admin@magma.test password1234
-docker-compose exec magmalte yarn setAdminPassword master admin@magma.test password1234
+docker-compose exec magmalte yarn setAdminPassword host admin@magma.test password1234
 
 # Docker run in a Linux host doesn't resolve host.docker.internal to the host IP.
 # See https://github.com/docker/for-linux/issues/264
