+++ nms/packages/magmalte/views/host.pug
@@ -24,4 +24,4 @@ html
       var CONFIG = !{configJson};
   body
     div(id="root")
-    script(src=staticDist("master.js", "nms"))
+    script(src=staticDist("host.js", "nms"))
